+++ app/Http/Controllers/Admin/ChambreController.php
@@ -7,7 +7,7 @@ use Illuminate\Http\Request;
 use App\Model\Admin\Immeuble;
 use MercurySeries\Flashy\Flashy;
 use App\Http\Controllers\Controller;
-
+use Illuminate\Support\Facades\Auth;
 class ChambreController extends Controller
 {
     /**
@@ -23,8 +23,13 @@ class ChambreController extends Controller
     
     public function index()
     {
-        $immeuble = Immeuble::all();
-        return view('admin.chambre.add',compact('immeuble'));
+        if (Auth::guard('admin')->user()->can('logement.create')) 
+        {
+            $immeuble = Immeuble::all();
+            return view('admin.chambre.add',compact('immeuble'));
+        }
+                
+        return redirect(route('admin.home'));
     }
 
     /**
@@ -45,21 +50,26 @@ class ChambreController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request,[
-            'name' => 'required',
-            'immeuble' => 'required',
-            'nombre' => 'required|numeric',
-            'genre' => 'required',
-        ]);
-            $add_chambre = new Chambre;
-            $add_chambre->nom = $request->name;
-            $add_chambre->nombre = $request->nombre;
-            $add_chambre->status = 1;
-            $add_chambre->genre = $request->genre;
-            $add_chambre->save();
-            $add_chambre->immeubles()->sync($request->immeuble);
-            Flashy::success('Votre chambre a ete ajoute');
-            return redirect()->route('admin.logement.index');
+        if (Auth::guard('admin')->user()->can('logement.create')) 
+        {
+            $this->validate($request,[
+                'name' => 'required',
+                'immeuble' => 'required',
+                'nombre' => 'required|numeric',
+                'genre' => 'required',
+            ]);
+                $add_chambre = new Chambre;
+                $add_chambre->nom = $request->name;
+                $add_chambre->nombre = $request->nombre;
+                $add_chambre->status = 1;
+                $add_chambre->genre = $request->genre;
+                $add_chambre->save();
+                $add_chambre->immeubles()->sync($request->immeuble);
+                Flashy::success('Votre chambre a ete ajoute');
+                return redirect()->route('admin.logement.index');
+        }
+                
+        return redirect(route('admin.home'));
     }
 
     /**
@@ -93,21 +103,26 @@ class ChambreController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request,[
-            'name' => 'required',
-            'nombre' => 'required|numeric',
-            'genre' => 'required',
-        ]);
-            $statusNumber = '';
-            $update_chambre = Chambre::find($id);
-            $update_chambre->nom = $request->name;
-            $update_chambre->nombre = $request->nombre;
-            $update_chambre->status = $request->status;
-            $update_chambre->genre = $request->genre;
-            $update_chambre->save();
-            $update_chambre->immeubles()->sync($request->immeuble);
-            Flashy::success('Votre chambre a ete ajoute');
-            return redirect()->route('admin.logement.index');
+        if (Auth::guard('admin')->user()->can('logement.update')) 
+        {
+            $this->validate($request,[
+                'name' => 'required',
+                'nombre' => 'required|numeric',
+                'genre' => 'required',
+            ]);
+                $statusNumber = '';
+                $update_chambre = Chambre::find($id);
+                $update_chambre->nom = $request->name;
+                $update_chambre->nombre = $request->nombre;
+                $update_chambre->status = $request->status;
+                $update_chambre->genre = $request->genre;
+                $update_chambre->save();
+                $update_chambre->immeubles()->sync($request->immeuble);
+                Flashy::success('Votre chambre a ete ajoute');
+                return redirect()->route('admin.logement.index');
+            }
+                    
+            return redirect(route('admin.home'));
     }
 
     /**
@@ -118,8 +133,13 @@ class ChambreController extends Controller
      */
     public function destroy($id)
     {
-        Chambre::find($id)->delete();
-        Flashy::success('Votre Chambre a ete Supprimer');
-        return back();
+        if (Auth::guard('admin')->user()->can('logement.delete')) 
+        {
+            Chambre::find($id)->delete();
+            Flashy::success('Votre Chambre a ete Supprimer');
+            return back();
+        }
+                        
+        return redirect(route('admin.home'));
     }
 }
